host/requirements.txt
altgraph==0.17.4
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pefile==2023.2.7
pillow==10.4.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pystray==0.19.5
pywin32-ctypes==0.2.2
setuptools==70.2.0
six==1.16.0
Werkzeug==3.0.3
